todo: performance considerations
The CPU profile showed that the highest overhead is due to object
creation, detection, and manipulation. It doesn't seem to have any
simple solution. However, an obvious improvement would be to
minimize any data transformation during the most critical part - data
updating process. In the future it might be possible to refactor the
program to do as minimal interactions during price updates as possible.
This would minimize the overhead and guarantee fast execution time until
order execution. However currently it is not a viable option, due to
program being at early stage of development. Therefore, the current
structure will be kept until it will be reasonably possible to optimize.